feat(toggled reaction roles!)

+++ src/reactionRoleManager.js
@@ -8,14 +8,15 @@ const sleep = (ms) => new Promise(resolve => setTimeout(resolve, ms));
 * Structure from reaction role object.
 */
 class ReactionRole {
-    constructor({ message, channel, guild, role, emoji, winners, max }) {
+    constructor({ message, channel, guild, role, emoji, winners, max, toggle }) {
         this.message = message.id ? message.id : message;
         this.channel = message.channel ? message.channel.id : channel;
         this.guild = message.guild ? message.guild.id : guild;
         this.role = role.id ? role.id : role;
         this.emoji = emoji.id ? emoji.id : emoji.name || emoji;
         this.winners = winners || [];
-        this.max = max != null && !isNaN(Number(max)) ? max : Number.MAX_SAFE_INTEGER;
+        this.max = isNaN(max) ? Number.MAX_SAFE_INTEGER : max;
+        this.toggle = Boolean(toggle);
     }
 
     get id() {
@@ -24,7 +25,15 @@ class ReactionRole {
 
     toJSON() {
         return {
-            id: this.id, message: this.message, channel: this.channel, guild: this.guild, role: this.role, emoji: this.emoji, winners: this.winners, max: this.max > Number.MAX_SAFE_INTEGER ? Number.MAX_SAFE_INTEGER : this.max
+            id: this.id,
+            message: this.message,
+            channel: this.channel,
+            guild: this.guild,
+            role: this.role,
+            emoji: this.emoji,
+            winners: this.winners,
+            max: this.max > Number.MAX_SAFE_INTEGER ? Number.MAX_SAFE_INTEGER : this.max,
+            toggle: this.toggle
         };
     }
 
@@ -36,7 +45,8 @@ class ReactionRole {
             role: json.role,
             emoji: json.emoji,
             winners: json.winners,
-            max: json.max
+            max: json.max,
+            toggle: json.toggle
         })
     }
 }
@@ -80,8 +90,16 @@ class ReactionRoleManager extends EventEmitter {
                 role: String,
                 emoji: String,
                 winners: Array,
-                max: Number
+                max: {
+                    type: Number,
+                    default: Number.MAX_SAFE_INTEGER
+                },
+                toggle: {
+                    type: Boolean,
+                    default: false
+                }
             }));
+            return Promise.resolve();
         } catch (e) {
             return Promise.reject(e);
         }
@@ -92,10 +110,11 @@ class ReactionRoleManager extends EventEmitter {
             return;
 
         await this.__checkMongoose().catch(console.error);
-        await sleep(1000);
         await this.__parseStorage();
+        await sleep(1000);
 
         const updateQueue = [];
+        const checkedRoles = [];
         for (const reactionRole of this.roles.values()) {
             const guild = this.client.guilds.cache.get(reactionRole.guild);
             if (!guild) {
@@ -103,7 +122,6 @@ class ReactionRoleManager extends EventEmitter {
                 this.deleteReactionRole(reactionRole);
                 continue;
             }
-
 
             const channel = guild.channels.cache.get(reactionRole.channel);
             if (!channel) {
@@ -122,57 +140,84 @@ class ReactionRoleManager extends EventEmitter {
             if (!message.reactions.cache.has(reactionRole.emoji)) // Bot reaction is deleted, and not have any reaction with this reaction role emoji.
                 await message.react(reactionRole.emoji);
 
-            for (const reaction of message.reactions.cache.values()) {
-                await reaction.users.fetch(); //Need fetch the users to next for
-                for (const user of reaction.users.cache.values()) {
-                    if (user.bot) // Ignore bots, please!
-                        continue;
+            // for (const reaction of message.reactions.cache.values()) {
+            const reaction = message.reactions.cache.find(x => `${message.id}-${this.client.emojis.resolveIdentifier(x.emoji.id || x.emoji.name)}`);
+            /* if (!reaction) {
+                 console.log('not found! wtf');
+                 continue;
+             }*/
 
-                    const member = guild.members.cache.get(user.id);
-                    if (!member) {
-                        await reaction.remove(user);
-                        this.__debug('REACTION', `Member '${user.id}' wasn't found, reaction of his was removed from message.`);
-                        continue;
-                    }
+            await reaction.users.fetch(); //Need fetch the users to next for
+            for (const user of reaction.users.cache.values()) {
+                if (user.bot) // Ignore bots, please!
+                    continue;
 
-                    const emoji = this.client.emojis.resolveIdentifier(reaction.emoji.id || reaction.emoji.name);
-                    const id = `${message.id}-${emoji}`;
-                    if (id != reactionRole.id) {
-                        this.__debug('ROLE', `Weird problem, some reaction with wrong setup. Id '${id}' from this role is not '${reactionRole.id}'...`);
-                        continue;
-                    }
-
-                    if (reactionRole.winners.indexOf(member.id) <= -1)
-                        reactionRole.winners.push(member.id);
-                    if (!member.roles.cache.has(reactionRole.role)) {
-                        this.emit('reactionRoleAdd', member, reactionRole.role);
-                        await member.roles.add(reactionRole.role);
-                        this.__debug('ROLE', `Role '${reactionRole.role}' was given to '${member.id}', it reacted when bot wasn't online.`);
-                    }
-                    else {
-                        this.__debug('ROLE', `Keeping role '${reactionRole.role}' from '${member.id}', it reacted and already have the role.`);
-                    }
+                const member = guild.members.cache.get(user.id);
+                if (!member) {
+                    await reaction.users.remove(user);
+                    this.__debug('REACTION', `Member '${user.id}' wasn't found, reaction of his was removed from message.`);
+                    continue;
                 }
 
-                for (const winnerId of reactionRole.winners) {
-                    const member = guild.members.cache.get(winnerId);
-                    if (!member) {
-                        const index = reactionRole.winners.indexOf(winnerId);
-                        reactionRole.winners.splice(index, 1);
-                        this.__debug('BOOT', `Member '${winnerId}' wasn't found, his was removed from winner list.`);
-                        continue;
-                    }
-
-                    if (!reaction.users.cache.has(winnerId)) {
-                        if (member.roles.cache.has(reactionRole.role)) {
-                            await member.roles.remove(reactionRole.role);
-                            this.emit('reactionRoleRemove', member, reactionRole.role);
+                /* const emoji = this.client.emojis.resolveIdentifier(reaction.emoji.id || reaction.emoji.name);
+                 const id = `${message.id}-${emoji}`;
+                 if (id != reactionRole.id) {
+                     this.__debug('ROLE', `Weird problem, some reaction with wrong setup. Id '${id}' from this role is not '${reactionRole.id}'...`);
+                     continue;
+                 }
+*/
+                if (reactionRole.toggle) {
+                    const uncheckedRoles = this.roles.filter(role => role.message == reactionRole.message).filter(x => !checkedRoles.includes(`${x.role}-${member.id}`));
+                    let counter = 0;
+                    for (const [_, unchecked] of uncheckedRoles) {
+                        if (counter >= 1 && !checkedRoles.includes(`${unchecked.role}-${member.id}`)) {
+                            checkedRoles.push(`${unchecked.role}-${member.id}`);
+                            await reaction.users.remove(user);
+                            if (member.roles.cache.has(unchecked.role)) {
+                                await member.roles.remove(unchecked.role);
+                                this.emit('reactionRoleRemove', member, reactionRole.role);
+                            }
                         }
-
-                        const index = reactionRole.winners.indexOf(winnerId);
-                        reactionRole.winners.splice(index, 1);
-                        this.__debug('ROLE', `Role '${reactionRole.role}' removed from '${member.id}', it removed reaction when bot wasn't online.`)
+                        counter++;
                     }
+                    if (counter >= 1)
+                        continue;
+                }
+
+                if (reactionRole.winners.indexOf(member.id) <= -1)
+                    reactionRole.winners.push(member.id);
+                if (!member.roles.cache.has(reactionRole.role)) {
+                    this.emit('reactionRoleAdd', member, reactionRole.role);
+                    await member.roles.add(reactionRole.role);
+                    this.__debug('ROLE', `Role '${reactionRole.role}' was given to '${member.id}', it reacted when bot wasn't online.`);
+                }
+                else {
+                    this.__debug('ROLE', `Keeping role '${reactionRole.role}' from '${member.id}', it reacted and already have the role.`);
+                }
+            }
+            //  }
+
+            for (const winnerId of reactionRole.winners) {
+                const member = guild.members.cache.get(winnerId);
+                if (!member) {
+                    const index = reactionRole.winners.indexOf(winnerId);
+                    reactionRole.winners.splice(index, 1);
+                    this.__debug('BOOT', `Member '${winnerId}' wasn't found, his was removed from winner list.`);
+                    continue;
+                }
+
+                if (member.user.bot)
+                    continue;
+
+                if (!reaction.users.cache.has(winnerId)) { // Delete role if user reacted off
+                    if (member.roles.cache.has(reactionRole.role)) {
+                        await member.roles.remove(reactionRole.role);
+                        this.emit('reactionRoleRemove', member, reactionRole.role);
+                    }
+
+                    const index = reactionRole.winners.indexOf(winnerId);
+                    reactionRole.winners.splice(index, 1);
+                    this.__debug('ROLE', `Role '${reactionRole.role}' removed from '${member.id}', it removed reaction when bot wasn't online.`)
                 }
             }
             updateQueue.push(reactionRole);
@@ -192,8 +237,9 @@ class ReactionRoleManager extends EventEmitter {
     * @property {Role} options.role - Role what the bot will give/take from members when they react.
     * @property {Emoji} options.emoji - Emoji or emoji id what member will react to win/lose the role.
     * @property {Number} [options.max] - Max roles to give, default is Infinity.
+    * @property {Boolean} [options.toggle] - User will have only one of these message roles, default false.
     */
-    createReactionRole({ message, role, emoji, max } = { max: Number.MAX_SAFE_INTEGER }) {
+    createReactionRole({ message, role, emoji, max, toggle } = { max: Number.MAX_SAFE_INTEGER, toggle: false }) {
         return new Promise(async (resolve, reject) => {
             if (message instanceof Message) {
                 if (!message.guild)
@@ -207,9 +253,12 @@ class ReactionRoleManager extends EventEmitter {
                     return reject('Bad input: I canno\'t resolve emoji ' + role);
 
                 await message.react(emoji);
-                const reactionRole = new ReactionRole({ message: message, role, emoji, max });
+                const reactionRole = new ReactionRole({ message: message, role, emoji, max, toggle });
                 this.roles.set(reactionRole.id, reactionRole);
-                this.__store(reactionRole);
+                await this.__store(reactionRole);
+                if (toggle) {
+                    await this.mongoose.model('ReactionRoles').updateMany({ message: message }, { toggle: toggle }).exec();
+                }
                 this.__debug('ROLE', `Role '${role.id}' added in reactionRoleManager!`);
                 return resolve();
             }
@@ -256,7 +305,8 @@ class ReactionRoleManager extends EventEmitter {
                     await this.mongoose.model('ReactionRoles').findOneAndUpdate({ id: role.id }, role, { new: true, upsert: true }).exec();
                 }
             }
-            else {
+
+            if (fs.existsSync(this.DATA_JSON_PATH)) {
                 fs.writeFileSync(this.DATA_JSON_PATH, JSON.stringify(this.roles.map(role => role.toJSON())));
                 this.__debug('STORE', `Stored roles saved, contains '${this.roles.size}' roles.`);
             }
@@ -265,22 +315,21 @@ class ReactionRoleManager extends EventEmitter {
 
     async __parseStorage() {
         if (this.storage) {
-            if (this.mongoose) {
-                const reactionRoles = await this.mongoose.model('ReactionRoles').find({});
-                for (const role of reactionRoles) {
-                    this.roles.set(role.id, new ReactionRole(role));
-                }
+            const roles = [];
+            if (fs.existsSync(this.DATA_JSON_PATH)) {
+                const json = JSON.parse(fs.readFileSync(this.DATA_JSON_PATH).toString());
+                roles.push(...json);
             }
-            else {
-                if (fs.existsSync(this.DATA_JSON_PATH)) {
-                    const json = JSON.parse(fs.readFileSync(this.DATA_JSON_PATH).toString());
-                    for (const role of json) {
-                        this.roles.set(role.id, ReactionRole.fromJSON(role));
-                    }
-                    this.__debug('STORE', `Stored roles parsed, contains '${roles.size}' roles.`);
-                }
+
+            if (this.mongoose) {
+                roles.push(...await this.mongoose.model('ReactionRoles').find({}));
+            }
+
+            for (const role of roles) {
+                this.roles.set(role.id, ReactionRole.fromJSON(role));
             }
         }
+        this.__debug('STORE', `Stored roles parsed, contains '${this.roles.size}' roles.`);
     }
 
     async __onReactionAdd(msgReaction, user) {
@@ -314,6 +363,28 @@ class ReactionRoleManager extends EventEmitter {
         if (reactionRole.winners.indexOf(member.id) <= -1)
             reactionRole.winners.push(member.id);
 
+        if (reactionRole.toggle) {
+            for (const reaction of message.reactions.cache.values()) {
+                await reaction.users.fetch();
+                if (reaction.users.cache.has(user.id)) {
+                    const emoji = this.client.emojis.resolveIdentifier(reaction.emoji.id || reaction.emoji.name);
+                    const reactionId = `${message.id}-${emoji}`;
+                    const rr = this.roles.get(reactionId);
+                    if (id == reactionId)
+                        continue;
+
+                    await reaction.users.remove(user);
+                    if (rr && member.roles.cache.has(rr.role)) {
+                        await member.roles.remove(rr.role);
+                        this.emit('reactionRoleRemove', member, role);
+                    }
+
+                    if (rr.winners.indexOf(member.id) <= -1)
+                        rr.winners.push(member.id);
+                }
+            }
+        }
+
         await member.roles.add(role).catch(console.error);
         this.emit('reactionRoleAdd', member, role);
         this.__debug('REACTION', `User '${member.displayName}' won the role '${role.name}'.`);
@@ -339,14 +410,14 @@ class ReactionRoleManager extends EventEmitter {
 
         const role = guild.roles.cache.get(reactionRole.role);
         if (!(role instanceof Role)) {
-            this.deleteReactionRole(reactionRole);
             this.__debug('ROLE', `Role '${reactionRole.role}' wasn't found in guild '${guild.id}', the member '${member.id}' will not lose the role.`)
+            return this.deleteReactionRole(reactionRole);
         }
 
         const index = reactionRole.winners.indexOf(member.id);
         reactionRole.winners.splice(index, 1);
 
-        if (member.roles.cache.has(role)) {
+        if (member.roles.cache.has(role.id)) {
             await member.roles.remove(role).catch(console.error);
             this.emit('reactionRoleRemove', member, role);
             this.__debug('REACTION', `User '${member.displayName}' lost the role '${role.name}'.`);
